requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
faiss-cpu
groq
